interactive test to exercise integrity functionality.
This commit adds an interactive test (i.e. a human needs to look at
the output) to start exercising the functionality for integrity. An
initial test endpoint that simulates a diagnostic test on another
service has been added to avoid dependencies on the full bc ecosystem.
This should keep the service light and lets me actually iterate while
on an airplane with spotty wifi.