define targets and tests separately.
The goal is to be able to run integrity tests against a canary
"pool" (this could be every store on the platform or just a subset)
and to make this easier, this commit does an outer product of the
target pool and the test suite. Moving this into SQL should prove
useful.