Simplify resource integrity checks.
My original idea was to maximize flexiblity - storing a complete list
of paths to various resources across a distribute architecture for
each member of the test pool. This would let ids vary or container
multiple elements.

However, that is fairly complicated. I don't like complicated. This
commit implements an approach that imposes the following:

 - An integrity test runs on a single resource like a store, user,
   product, integration, etc.

 - The resource is assumed to have the same identifier across all
   systems. system1/products/1234 and system2/products/1234.

 - The resource is assumed to be accessible via the same path on a
   given system. /products/:id for all instances of the resource.

This makes it very easy to bundle together a list of resources and
tests, but does impose some conventions on how diagnostic endpoints
are implemented. This commit decides in favor of convention over
configuration when handling the trade-off between approaches.