Build out data structure for running tests.
This commit bundles targets and tests into panels that can be named
and run. Eventually, I imagine these panels will have their own
dashboards and such. One way to think about this is that a panel has
a designated group of canaries and a set of tests to execute against
them.