Describe tasks in files.
The problem I'm trying to tackle is how to reliably tell the system
what to do, given the following use cases:

 - A support engineer might want to quickly diagnose why a resource
   is not working.

 - A support engineer might want to catch errors early and need to
   poll resources regularly.

 - A data analyst might want to figure out how widespread an error
   is across many resources.

There are a lot of ways to attack task definition, but the one that
seems best to me right now is freeform. We accept the full task def
in a POST request body (or file as is the case here) and keep a copy
around. The copy can be re-run, put on a recurring schedule and
modified over time. This is basically how RunDeck, Amazon ECS and
Heroku all work.

Besides being fairly simple, the other major advantage is data
integrity- the task contains all the data required to do that job. A
MySQL based solution would risk task definitions changing during
test runs.

To that end, this commit moves the task definition out of go and into
a flat file that can also be curled, should I build the endpoint.