Send test results to worker after request.
I've never tried this before. Each test case includes a channel that
is consumed by the original task job. As a test case is processed by
a retriever, the results are sent back to the task for processing.
This should allow request retrieval to be decoupled into a job/worker
system.