Serve latest test results.
This commit keeps a copy of the latest test results for each task
and serves them on a web page. As results become available, they are
packed into an in-memory map. While this sacrifices durability, it
does increase simplicity and reduce dependencies. The full reasoning
is detailed in the README.