Harness up Task.
This commit gets Task into a test harness that executes an arbitrary
task and processes fake test results. Interestingly, the shared
mutable state on integrity creates some issues. The immediate one is
that the test doesn't block until all the data has been processed,
so the mutable state is empty.

Do we solve this with Mutexes or Channels?

I understand how to solve it with channels - the test just blocks
until a new result is passed out to the big result map and I don't
understand how to use mutexes in an efficient way to block the test.